+++ internal/state/states.go
@@ -1,5 +1,7 @@
 package state
 
+import "time"
+
 type States interface {
 	// GetState Puts a new user into the context. If the user already exists in the context, it's kept,
 	GetState(userId int64) (*UserState, error)
@@ -7,13 +9,14 @@ type States interface {
 }
 
 type UserState struct {
-	UserId                int64    `firestore:"userId"`
-	CurrentCategoryNodeId string   `firestore:"currentCategoryNodeId"`
-	MessageText           string   `firestore:"messageText"`
-	Files                 []string `firestore:"files"`
-	Login                 string   `firestore:"login"`
-	Password              string   `firestore:"password"`
-	Token                 string   `firestore:"token"`
-	MessageHandlerName    string   `firestore:"messageHandlerName"`
-	SentMessagesCount     int      `firestore:"sentMessagesCount"`
+	UserId                int64     `firestore:"userId"`
+	CurrentCategoryNodeId string    `firestore:"currentCategoryNodeId"`
+	MessageText           string    `firestore:"messageText"`
+	Files                 []string  `firestore:"files"`
+	Login                 string    `firestore:"login"`
+	Password              string    `firestore:"password"`
+	Token                 string    `firestore:"token"`
+	RateLimitedUntil      time.Time `firestore:"rateLimitedUntil"`
+	MessageHandlerName    string    `firestore:"messageHandlerName"`
+	SentMessagesCount     int       `firestore:"sentMessagesCount"`
 }