+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Do not try sending a message if user is rate limited 
+
 ## [0.7.1] - 2023-03-12
 
 ### Fixed