+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Show if user is authorized on the portal or not
+- Store username on `/start` command
+
 ## [0.8.0] - 2023-03-18
 
 ### Added