+++ CHANGELOG.md
@@ -7,11 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9.2] - 2023-03-20
+
 ### Fixed
 
 - Endless loop when can't download files or another error
 
-## [0.9.1] - 2023-03-18
+## [0.9.1] - 2023-03-19
 
 ### Fixed